src/config.py
"""加载 YAML 配置为字典。"""
from pathlib import Path
from typing import Any

import yaml


def load_config(path: str | Path) -> dict[str, Any]:
    path = Path(path)
    with path.open(encoding="utf-8") as f:
        return yaml.safe_load(f)
