requirements.txt
# 深度学习核心
torch>=2.0.0
torchvision>=0.15.0

# 配置与实验
pyyaml>=6.0

# 可视化与进度条
matplotlib>=3.7.0
tqdm>=4.65.0

# Notebook（可选）
jupyter>=1.0.0
ipykernel>=6.25.0
